[refactor] 페이지 관련 파일 위치 변경
