+++ package.json
@@ -11,11 +11,14 @@
   "dependencies": {
     "@supabase/ssr": "^0.6.1",
     "@supabase/supabase-js": "^2.50.0",
+    "date-fns": "^4.1.0",
     "mysql": "^2.18.1",
     "mysql2": "^3.14.1",
     "next": "15.3.3",
     "react": "^19.0.0",
+    "react-date-range": "^2.0.1",
     "react-dom": "^19.0.0",
+    "react-flags-select": "^2.5.0",
     "uuid": "^11.1.0"
   },
   "devDependencies": {
